major: removed external api call
